requirements.txt
# Requirements.txt content:
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
jupyter==1.0.0
fake-useragent==1.4.0
selenium==4.15.0
lxml==4.9.3
python-dotenv==1.0.0
tqdm==4.66.1